+++ Modular Gameplay Overhaul/Modules/Core/CHANGELOG.md
@@ -1,5 +1,11 @@
 ﻿# Modular Overhaul Core Change Log
 
+## 2.0.5
+
+### Fixed
+
+* Default value for Combat module has been correctly set to `false`. This would have prevented the initial setup wizard from triggering.
+
 ## 2.0.1
 
 ### Fixed